+++ Content.Client/Interfaces/IItemSlotManager.cs
@@ -1,4 +1,4 @@
-﻿using Robust.Client.UserInterface.Controls;
+﻿using Robust.Client.UserInterface;
 using Robust.Shared.Interfaces.GameObjects;
 
 namespace Content.Client.Interfaces
@@ -6,6 +6,6 @@ namespace Content.Client.Interfaces
     public interface IItemSlotManager
     {
         public void Initialize();
-        public bool OnButtonPressed(BaseButton.ButtonEventArgs args, IEntity item);
+        public bool OnButtonPressed(GUIBoundKeyEventArgs args, IEntity item);
     }
 }
